+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixes a bug in the windowManager where positions could not be given as a number string
 - Add export of `replaceAttributes` which inserts attributes into a template string.
+- Fixes a bug in app `getState` where empty plugin state objects where added to the app state
 
 # 5.2